% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bica}
\alias{bica-package}
\alias{bica}
\title{bica: What the Package Does (One Line, Title Case)}
\description{
Bayesian inseason Chinook abundance model.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\concept{help}
