% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_day_func}
\alias{my_day_func}
\title{Getting day of year}
\usage{
my_day_func(Month, Day)
}
\arguments{
\item{Month}{Numeric month}

\item{Day}{Numeric day}
}
\value{
Julian date (non-leap year) for the Inseason Bayesian Projection Model
}
\description{
Getting day of year
}
\examples{
my_day_func(6, 10)

}
