% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{run_bica_model}
\alias{run_bica_model}
\title{Format data for BICA model}
\usage{
run_bica_model(
  bica_data,
  model_version,
  normal = FALSE,
  logistic = FALSE,
  n_chains,
  n_iter,
  n_thin
)
}
\arguments{
\item{bica_data}{Input data as a list (from format _bica_data())}

\item{model_version}{Character string. Model version to run.}

\item{normal}{Logical, \code{TRUE} if running "pss_norm..." model}

\item{logistic}{Logical, \code{TRUE} if running "pss_logistic..." model}

\item{n_chains}{Number of MCMC chains}

\item{n_iter}{Number of MCMC iterations}

\item{n_thin}{Amount of thinning for MCMC chains}
}
\description{
Format data for BICA model
}
