% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tr_plot}
\alias{tr_plot}
\title{Make trace plots}
\usage{
tr_plot(obj, vars)
}
\arguments{
\item{obj}{Posterior trace from model output}

\item{vars}{Variables being plotted}
}
\value{
Trace plots of parallel MCMC chains in ggplot
}
\description{
Make trace plots
}
\examples{
\dontrun{
tr_plot(bica_out$pars, c("alpha", "beta", "sigma", "ln_run_size", "lp__"))
}

}
