% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format.R
\name{format_bica_data}
\alias{format_bica_data}
\title{Format data for BICA model}
\usage{
format_bica_data(
  my_year,
  my_day,
  end_year,
  pf_hist,
  can_hist,
  pss_hist,
  eagle_hist,
  gsi_by_year,
  pss_sd,
  prior_df_log,
  prior_df_norm,
  start_day_pss = 148,
  start_year_pss = 1995
)
}
\arguments{
\item{my_year}{Year the projection is made}

\item{my_day}{Day the projection is made}

\item{end_year}{Last Complete Year}

\item{pf_hist}{PF for 2007-present}

\item{can_hist}{EOS total Canadian-origin Chinook est (DF)}

\item{pss_hist}{Daily PSS passage (DF)}

\item{eagle_hist}{Eagle sonar passage}

\item{gsi_by_year}{Genetic apportionment of Can-origin Chinook}

\item{pss_sd}{Variance estimates for PSS}

\item{prior_df_log}{priors for logistic params}

\item{prior_df_norm}{priors for normal params}

\item{start_day_pss}{First day of PSS calculation}

\item{start_year_pss}{First year to include in PSS}
}
\description{
Format data for BICA model
}
