% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{run_bica_model}
\alias{run_bica_model}
\title{Format data for BICA model}
\usage{
run_bica_model(bica_data, model_version, n_chains, n_iter, n_thin)
}
\arguments{
\item{bica_data}{Input data as a list (from format_bica_data())}

\item{model_version}{Character string. Model version to run.}

\item{n_chains}{Number of MCMC chains}

\item{n_iter}{Number of MCMC iterations}

\item{n_thin}{Amount of thinning for MCMC chains}
}
\value{
Results and summaries from BICA model run
}
\description{
Format data for BICA model
}
\examples{
\donotrun{
bica_out <- bica::run_bica_model(bica_data, model_version = "pss_prop_es_prop", n_chains = 4, n_iter = 5000, n_thin = 5)

bica_out$summary
}

}
